+++ CHANGES.md
@@ -3,6 +3,7 @@ Change Log
 
 ### Next Release
 
+* If `showIEMessage` is `true` in config.json, warn IE11 users that support is ending.
 
 ### v7.11.10